streamlit-llm_LangChain/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
langchain-community
anthropic
langchain-openai
langchain-anthropic
langchain-google-genai
faiss-cpu
